change parse int to float for dron_speed

+++ virtualDron_anomaly.py
@@ -47,7 +47,7 @@ complete_url = base_url + "appid=" + api_key + "&units=metric"
 mqtt_address            = os.environ.get('MQTT_ADDRESS')
 mqtt_port               = int(os.environ.get('MQTT_PORT'))
 num_drones              = int(os.environ.get('NUM_DRONES'))
-dron_speed              = int(os.environ.get('DRON_SPEED'))
+dron_speed              = float(os.environ.get('DRON_SPEED'))
 wait_client_seconds     = int(os.environ.get('WAIT_CLIENT_SECONDS'))
 delta_time = 0.33
 # ------------------------------------------------------------------------------ #
